model_training/requirements.txt
torch
torchvision
albumentations
pandas
numpy
scikit-learn
Pillow
grad-cam
matplotlib
seaborn
opencv-python
tqdm
